config: add mathjax settings
load MathJax 2.7 from a CDN, and configure it to recognize TeX in-line delimiters `$...$`

Note:
Kramdown expects to see mathematical formulas delimited by `$$...$$`, and then generates the corresponding html as `<script type="math/tex>...</script>` or `<script type="math/tex"; mode=display>...</script>`. This math/tex tag works in MathJax 2 but not in MathJax3. Further modifications to revert the tag back to display delimiters (e.g. `\\[...\\]`) are needed if we want to use MathJax3.

Ref:
1. https://11011110.github.io/blog/2019/10/17/mathjax-3-jekyll.html
2. http://benlansdell.github.io/computing/mathjax/